RightStufAnime Scrap Additions
Fixed issue where the url is valid but no data is scraped due to the soap scrape method being called before the webpage is fully loaded. Added functionality that asks the user whether they want the data for a manga or a light novel and then the title of the series. Cleaned up some stuff and added comments